Rename endian.h into efivar_endian.h to avoid conflicts
When compiling efivar from another directory than the sources's one, we
will have to add ./src/ in the include path to be able to resolve
the include "endian.h" from ./src/dp-message.c. Then, when include
<endian.h> is done from ./src/guid.h, ./src/endian.h will be used
instead of system wide endian.h, which is wrong.

To avoid this situation, this patch renames ./src/endian.h into
./src/efivar_endian.h
