+++ src/efivar_endian.h
